Move the test to experimental rekt
