+++ test/experimental/features/eventtype_autocreation/feature.go
@@ -1,23 +1,8 @@
-/*
-Copyright 2022 The Knative Authors
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package eventtype
+package eventtype_autocreation
 
 import (
 	"context"
+	"embed"
 	"time"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -28,7 +13,11 @@ import (
 	"knative.dev/reconciler-test/pkg/environment"
 	"knative.dev/reconciler-test/pkg/feature"
 	"knative.dev/reconciler-test/pkg/k8s"
+	"knative.dev/reconciler-test/pkg/manifest"
 )
+
+//go:embed eventtype.yaml
+var yaml embed.FS
 
 type EventType struct {
 	Name       string
@@ -73,4 +62,12 @@ func AssertPresent(expectedCeTypes sets.String) EventType {
 		},
 	}
 
+}
+
+// The function will apply the config map eventtype.yaml file to enable auto creation of eventtype
+// The yaml file is in /test/eventtype.yaml
+func ApplyEventTypeConfigMap() feature.StepFn {
+	return func(ctx context.Context, t feature.T) {
+		manifest.InstallYamlFS(ctx, yaml, nil)
+	}
 }
